Create reusable markdown component
